feat: add adonis typings
